Clean up legacy Discovery v1 server
Right now we have two discovery servers - v1 and v2. The v1 server
basically does nothing, except expose a single unused debugging
endpoint.

This PR kills the v1 discovery server, and moves the only part that is
used (setting up the profile endpoints) to its appropriate location in
InitDebug.
