+++ tools/hyperistio/hyperistio.go
@@ -29,7 +29,6 @@ import (
 
 	mixerEnv "istio.io/istio/mixer/test/client/env"
 	"istio.io/istio/pilot/pkg/bootstrap"
-	"istio.io/istio/pilot/pkg/proxy/envoy"
 	"istio.io/istio/pilot/pkg/serviceregistry"
 	envoyBootstrap "istio.io/istio/pkg/bootstrap"
 	"istio.io/istio/pkg/config/mesh"
@@ -159,11 +158,10 @@ func startPilot() error {
 	// Create a test pilot discovery service configured to watch the tempDir.
 	args := bootstrap.PilotArgs{
 		Namespace: "testing",
-		DiscoveryOptions: envoy.DiscoveryServiceOptions{
+		DiscoveryOptions: bootstrap.DiscoveryServiceOptions{
 			HTTPAddr:        ":15007",
 			GrpcAddr:        ":15010",
 			SecureGrpcAddr:  ":15011",
-			EnableCaching:   true,
 			EnableProfiling: true,
 		},
